Update nest_retention.py
Add some logging for the big errors. It was also hitting the ```if napi.authorization_required:``` section very frequently, so I have it trying to run the ```napi.request_token(pin)``` piece. No idea why it can work without giving it a pin, but I'm not one to shoot a gift horse.